requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coverage==4.5.3
entrypoints==0.3
file-magic==0.4.0
flake8==3.7.7
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-JSGlue==0.3.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10.1
jsonschema==3.0.1
lark-parser==0.7.1
lazy-object-proxy==1.3.1
lxml==4.3.3
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
git+https://github.com/bodik/nessus-report-parser.git@latest_cummulative#egg=nessus_report_parser
pathlib2==2.3.2
pluggy==0.7.1
psycopg2-binary==2.7.6.1
py==1.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.1.1
pyrsistent==0.14.11
pytest==3.8.2
pytest-base-url==1.4.1
pytest-flask==0.15.0
pytest-html==1.20.0
pytest-metadata==1.8.0
pytest-selenium==1.16.0
pytest-variables==1.7.1
python-dateutil==2.8.0
python-libnmap==0.7.0
PyYAML==5.1
requests==2.21.0
selenium==3.141.0
six==1.11.0
SQLAlchemy==1.3.3
sqlalchemy-datatables==1.2.0
git+https://github.com/bodik/sqlalchemy-filters@latest_cummulative#egg=sqlalchemy_filters
sqlalchemy-json==0.2.3
SQLAlchemy-Utils==0.33.11
typed-ast==1.1.0
urllib3==1.24.3
waitress==1.1.0
WebOb==1.8.2
WebTest==2.0.32
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
